+++ R/get_vals_switching.R
@@ -1,4 +1,4 @@
-#' @name switching_series
+#' @name get_vals_switching
 #' @title Compute values of transition function to separate regimes
 #' @description Function to estimate transition values, which uses a smooth transition function as
 #' used in Auerbach and Gorodnichenko (2012). The time series used in the transition function
@@ -6,7 +6,7 @@
 #' @param switching_data A numeric vector.
 #' @param specs A \link{list}() with inputs as in \link{lp_nl}().
 #' @return \item{fz}{A numeric vector with values from smooth transition function \eqn{F(z_{t-1})}.}
-#'
+#' @keywords internal
 #' @references
 #' Auerbach, A. J., and  Gorodnichenko Y. (2012). "Measuring the Output Responses to Fiscal Policy."
 #' \emph{American Economic Journal: Economic Policy}, 4 (2): 1-27.
@@ -19,15 +19,13 @@
 
 
 
-switching_series <- function(switching_data, specs){
-
- # switching_data <- specs$switching
+get_vals_switching <- function(switching_data, specs){
 
  # Decide whether to use HP filter.
   if(specs$hp_filter == 1){
 
   # Use HP-filter to decompose switching variable.
-   filter_results  <-   hp_filter_c(matrix(switching_data), specs$lambda)
+   filter_results  <-   hp_filter(matrix(switching_data), specs$lambda)
    gamma_fz        <-   specs$gamma
    z_0             <-   as.numeric(scale(filter_results[[1]], center = TRUE))
    fz              <-   exp((-1)*gamma_fz*z_0)/(1 + exp((-1)*gamma_fz*z_0))