+++ R/lp_nl.R
@@ -64,6 +64,7 @@
 #'\item{fz}{A vector containing the values of the transition function F(z_{t-1}).}
 #'
 #' @export
+#'
 #' @references
 #'
 #' Akaike, H. (1974). "A new look at the statistical model identification", \emph{IEEE Transactions on Automatic Control}, 19 (6): 716–723.
@@ -319,7 +320,7 @@ lp_nl <- function(data_set_df, lags_lin  = NULL, lags_nl        = NULL,  lags_cr
   x_lin      <- data_lin[[2]]
 
   # Construct shock matrix
-  d <- reduced_var(y_lin, x_lin, data_set_df, specs)
+  d <- get_mat_chol(y_lin, x_lin, data_set_df, specs)
 
   # Matrices to store irfs for each horizon
   irf_temp_s1_mean  <-  matrix(NaN, specs$endog, specs$hor + 1)
@@ -379,7 +380,7 @@ lp_nl <- function(data_set_df, lags_lin  = NULL, lags_nl        = NULL,  lags_cr
          for (k in 1:specs$endog){ # Accounts for the reactions of the endogenous variables
 
            # Estimate coefficients and newey west std.err
-           nw_results       <- lpirfs::newey_west_c(yy[, k], xx, h)
+           nw_results       <- lpirfs::newey_west(yy[, k], xx, h)
            b                <- nw_results[[1]]
            std_err          <- sqrt(diag(nw_results[[2]]))*specs$confint
 
@@ -448,8 +449,9 @@ lp_nl <- function(data_set_df, lags_lin  = NULL, lags_nl        = NULL,  lags_cr
             for (k in 1:specs$endog){ # Accounts for the reactions of the endogenous variables
 
              # Find optimal lag length and select matrices from lists accordingly
-              val_criterion   <- lpirfs::find_lag_c(y_lin, x_lin, lag_crit, h, k,
+              val_criterion   <- lpirfs::get_vals_lagcrit(y_lin, x_lin, lag_crit, h, k,
                                                         specs$max_lags)
+
               lag_choice      <- which.min(val_criterion)
 
               yy              <- y_nl[[lag_choice]][, k]
@@ -459,7 +461,7 @@ lp_nl <- function(data_set_df, lags_lin  = NULL, lags_nl        = NULL,  lags_cr
               xx              <- xx[1:(dim(xx)[1] - h + 1),]
 
              # Estimate parameters and newey west standard errors
-              nw_results      <- lpirfs::newey_west_c(yy, xx, h)
+              nw_results      <- lpirfs::newey_west(yy, xx, h)
               b               <- nw_results[[1]]
               std_err         <- sqrt(diag(nw_results[[2]]))